+++ crates/libs/reactor/src/core/mod.rs
@@ -1,53 +1,18 @@
-pub mod animation;
 pub mod backend;
-pub mod callback;
-pub mod component;
-pub mod component_element;
-pub mod context;
-pub mod custom;
-pub mod dispatcher;
 pub mod element;
-pub mod element_ext;
-pub mod error_boundary;
+pub mod engine;
 pub mod generated_bindings;
-pub mod geometry;
-pub mod into_elements;
-pub mod keyboard;
-pub mod modifiers;
-pub mod prop_binding;
-pub mod rc_fn;
+pub mod interaction;
 pub mod reconciler;
-pub mod render_context;
-pub mod render_host;
-pub mod resource;
-pub mod rich_text;
-pub mod templated_list;
-pub mod theme;
+pub mod style;
 pub mod widget;
 pub mod widgets;
 
-pub use animation::*;
 pub use backend::*;
-pub use callback::*;
-pub use component::*;
-pub use component_element::*;
-pub use context::*;
-pub use custom::*;
-pub use dispatcher::*;
 pub use element::*;
-pub use element_ext::*;
-pub use error_boundary::*;
-pub use geometry::*;
-pub use into_elements::*;
-pub use keyboard::*;
-pub use modifiers::*;
-pub use prop_binding::*;
+pub use engine::*;
+pub use interaction::*;
 pub use reconciler::*;
-pub use render_context::*;
-pub use render_host::*;
-pub use resource::*;
-pub use rich_text::*;
-pub use templated_list::*;
-pub use theme::*;
+pub use style::*;
 pub use widget::*;
 pub use widgets::*;
